+++ src/Margin/BrowserMarginProvider.cs
@@ -28,7 +28,7 @@ namespace MarkdownEditor
             if (!TextDocumentFactoryService.TryGetTextDocument(wpfTextViewHost.TextView.TextDataModel.DocumentBuffer, out document))
                 return null;
 
-            return new BrowserMargin(wpfTextViewHost.TextView, document);
+            return wpfTextViewHost.TextView.Properties.GetOrCreateSingletonProperty(() => new BrowserMargin(wpfTextViewHost.TextView, document));
         }
     }
 
@@ -53,7 +53,24 @@ namespace MarkdownEditor
             if (!TextDocumentFactoryService.TryGetTextDocument(wpfTextViewHost.TextView.TextDataModel.DocumentBuffer, out document))
                 return null;
 
-            return new BrowserMargin(wpfTextViewHost.TextView, document);
+            return wpfTextViewHost.TextView.Properties.GetOrCreateSingletonProperty(() => new BrowserMargin(wpfTextViewHost.TextView, document));
+        }
+    }
+
+    [Export(typeof(IWpfTextViewMarginProvider))]
+    [Name("LiveSyncFactory")]
+    [Order(Before = PredefinedMarginNames.HorizontalScrollBarContainer)]
+    [MarginContainer(PredefinedMarginNames.BottomRightCorner)]
+    [ContentType(MarkdownLanguage.LanguageName)]
+    [TextViewRole(PredefinedTextViewRoles.Debuggable)] // This is to prevent the margin from loading in the diff view
+    public class LiveSyncMarginBottomProvider : IWpfTextViewMarginProvider
+    {
+        public IWpfTextViewMargin CreateMargin(IWpfTextViewHost wpfTextViewHost, IWpfTextViewMargin marginContainer)
+        {
+            if (!MarkdownEditorPackage.Options.EnablePreviewWindow)
+                return null;
+
+            return wpfTextViewHost.TextView.Properties.GetOrCreateSingletonProperty(() => new LiveSyncMargin(wpfTextViewHost.TextView));
         }
     }
 }