+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Roadmap
 
+- [x] Toggle live sync on individual document
 - [ ] Format document/selection (#12)
 - [ ] Support for custom stylesheets
 - [ ] Generate .html files nested under the markdown file